refactor: move identity usability policy out of components
decrypt-message.ts needs the usable-identity and usable-bundle checks to keep
them inside the decrypt operation, which pointed a src/features module at
src/components for the first time. The module holds no React — it is a thin
policy over crypto/pq/suites — so it moves next to what it wraps and the arrow
disappears. Import paths only; no logic change.
