+++ src/features/decrypt-message.ts
@@ -8,7 +8,7 @@ import { decryptPqMessage } from "@/crypto/pq/decrypt-orchestrator"
 import type { PqCryptoClient } from "@/crypto/pq/worker-client"
 import { zeroize } from "@/crypto/pq/zeroize"
 import { getOrCreateVaultKey } from "@/crypto/vault/vault-key"
-import { isUsableBundle, isUsableIdentity } from "@/components/key-detail/identity-policy"
+import { isUsableBundle, isUsableIdentity } from "@/crypto/pq/identity-policy"
 import { recordReceipt, type ReceiptVerdict } from "@/features/receipt-cache"
 import { bytesToHex, bytesToUtf8 } from "@/lib/bytes"
 import { buildV2Payload } from "@/qr/payload-v2"
