+++ src/pages/encrypt-page.tsx
@@ -12,12 +12,12 @@ import {
 import { toast } from "sonner"
 import { AppError, toAppError } from "@/crypto/errors"
 import { groupSymmetricKeys } from "@/crypto/key-generation"
+import { isUsableBundle, isUsableIdentity } from "@/crypto/pq/identity-policy"
 import { ACTIVE_PROFILE, assertActiveSuite, resolveSuite } from "@/crypto/pq/suites"
 import { generateArtifactId, shortId } from "@/crypto/random"
 import { useSensitiveSession, useTransientClear } from "@/app/providers"
 import { AnimatedQrFrames } from "@/components/animated-qr-frames"
 import { DetailRow } from "@/components/detail-row"
-import { isUsableBundle, isUsableIdentity } from "@/components/key-detail/identity-policy"
 import { NoAutofocusDialogContent } from "@/components/no-autofocus-dialog-content"
 import { QrDisplay } from "@/components/qr-display"
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert"
